requirements.txt
setuptools<81
sqlalchemy>=2.0.0
alembic>=1.12.0
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
openai-whisper>=20231117
faiss-cpu>=1.7.4
pyttsx3>=2.90
sounddevice>=0.4.6
soundfile>=0.12.1
webrtcvad-wheels>=2.0.13
loguru>=0.7.0
pydantic>=2.0.0
pyyaml>=6.0
pillow>=10.0.0
pytest>=7.4.0
PyJWT>=2.8.0
passlib>=1.7.4
bcrypt>=4.0.1
httpx>=0.25.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.0
email-validator>=2.0.0
python-socketio>=5.10.0
slowapi>=0.1.9
numpy>=1.24.0
plyer>=2.1.0
pyinstaller>=6.0.0
